refactor: divide card components from `StorageHostSettingsPanel`
